add NftContentType

+++ types/tx_request.go
@@ -41,7 +41,7 @@ type CreateCollectionTxReq struct {
 type MintNftTxReq struct {
 	To                  string
 	NftCollectionId     int64
-	NftContentType      int8
+	NftContentType      int64
 	CreatorTreasuryRate int64
 	MetaData            string
 	MutableAttributes   string
